Extract functions into central functions module
